feat(feedboard): 컨벤션을 변경하였다
