+++ src/main/java/toy/subscribe/feedboard/model/FeedBoard.java
@@ -24,7 +24,7 @@ public class FeedBoard extends BaseEntity {
     private String guid;
 
     @Builder(access = AccessLevel.PUBLIC)
-    public FeedBoard(final Long id, final String company, final String imgPath, final String title, final String guid) {
+    private FeedBoard(final Long id, final String company, final String imgPath, final String title, final String guid) {
         this.id = id;
         this.company = company;
         this.imgPath = imgPath;
