+++ src/main/java/toy/subscribe/feedboard/service/FeedBoardProvideService.java
@@ -1,8 +1,26 @@
 package toy.subscribe.feedboard.service;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import toy.subscribe.configs.http.log.repository.HttpLogRepository;
 import toy.subscribe.configs.http.wrapper.HttpResponseWrapper;
+import toy.subscribe.feedboard.dto.FeedBoardResponseDto;
+import toy.subscribe.feedboard.repository.FeedBoardRepository;
 
-public interface FeedBoardProvideService {
-    HttpResponseWrapper provideFeedBoardWrapper(Pageable pageable, String company, String title);
+@Service
+@RequiredArgsConstructor
+public class FeedBoardProvideService {
+    private final FeedBoardRepository feedBoardRepository;
+    private final HttpLogRepository requestLogRepository;
+
+    @Transactional(readOnly = true)
+    public HttpResponseWrapper<FeedBoardResponseDto> provideFeedBoardWrapper(Pageable pageable, String company, String title) {
+        return HttpResponseWrapper.<FeedBoardResponseDto>builder()
+                                  .pages(feedBoardRepository.findPage(pageable, company, title))
+                                  .visitorsOfDay(requestLogRepository.findCumulativeVisitors())
+                                  .visitorsOfReduce(requestLogRepository.findDau())
+                                  .build();
+    }
 }
