--- src/main/java/toy/subscribe/feedboard/service/FeedBoardProvideServiceImpl.java
@@ -1,27 +1,0 @@
-package toy.subscribe.feedboard.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import toy.subscribe.configs.http.log.repository.HttpLogRepository;
-import toy.subscribe.configs.http.wrapper.HttpResponseWrapper;
-import toy.subscribe.feedboard.dto.FeedBoardResponseDto;
-import toy.subscribe.feedboard.repository.FeedBoardRepository;
-
-@Service
-@RequiredArgsConstructor
-public class FeedBoardProvideServiceImpl implements FeedBoardProvideService {
-    private final FeedBoardRepository feedBoardRepository;
-    private final HttpLogRepository requestLogRepository;
-
-    @Override
-    @Transactional(readOnly = true)
-    public HttpResponseWrapper<FeedBoardResponseDto> provideFeedBoardWrapper(Pageable pageable, String company, String title) {
-        return HttpResponseWrapper.<FeedBoardResponseDto>builder()
-                                  .pages(feedBoardRepository.findPage(pageable, company, title))
-                                  .visitorsOfDay(requestLogRepository.findCumulativeVisitors())
-                                  .visitorsOfReduce(requestLogRepository.findDau())
-                                  .build();
-    }
-}
