--- src/main/java/toy/subscribe/feedboard/service/SchedulingService.java
@@ -1,5 +1,0 @@
-package toy.subscribe.feedboard.service;
-
-public interface SchedulingService {
-    void allGroupFeedCollect();
-}
